fix: address PR review feedback for node pool selector
- Fix stale OpenAPI description referencing "live utilization" (now config-only)
- Add minNodes <= maxNodes cross-field validation in NodePoolConfiguration
- Remove @LogExecution from NodePoolProperties POJO (noisy on getters/setters)
- Move buildNodePoolAffinity from DeployableManifestGenerator to KnativeManifestGenerator (only consumer)
- Add missing minNodes field to docs/configuration.md, API contract, and data model spec
- Add trailing newline to Deployment.java
- Add test for minNodes > maxNodes validation
